// packages/luna-core/src/type.d.ts
export type DataAttributes = {
    [key: `data-${string}`]: string | number | boolean;
};
export type AriaAttributes = {
    [key: `aria-${string}`]: string | number | boolean;
};
export type DataSource = {
    url: string;
    body?: BodyInit | Record<string, unknown>;
    cache?: RequestCache;
    headers?: HeadersInit;
    method?: string;
    namespace?: string;
};
export type Source = {
    [key: string]: DataSource | Array<unknown>;
};
export type Definition = Record<string, unknown>;
export type Orderable = {
    order?: number;
};
export type Hideable = {
    hidden?: boolean;
};
// ...
export type Base = Orderable & Hideable;
export type CommonProps = {
    disabled?: boolean;
    id?: string;
    name?: string;
    placeholder?: string;
    required?: boolean;
};
// ...
export type CustomValidation = {
    field: string;
    message?: string;
    operator?: Operator;
};
export type FieldState = {
    disabled?: boolean;
    hidden?: boolean;
};
export type ActionEvent<T> = {
    action: T;
};
export type SourceEvent = ActionEvent<'source'> & {
    source: DataSource;
    target: string;
};
export type ValueEvent = ActionEvent<'value'> & {
    value: Record<string, Value | Array<Record<string, unknown>>>;
};
export type StateEvent = ActionEvent<'state'> & {
    state: FieldState;
    target: string | string[];
    when?: string | string[] | Condition;
};
export type ChangeEvent = Array<SourceEvent | ValueEvent | StateEvent>;
export type Validation = {
    custom?: CustomValidation | Array<CustomValidation>;
    email?: string;
    length?: Length<string>;
    required?: string;
};
export type Field = CommonProps & {
    advanced?: {
        aria?: AriaAttributes;
        autocomplete?: string;
        cols?: number;
        data?: DataAttributes;
        entity?: string;
        horizontal?: boolean;
        reverse?: boolean;
    };
    event?: {
        change?: ChangeEvent;
    };
    description?: string;
    fields?: never;
    label?: string;
    name: string;
    readonly?: boolean;
    type: string;
    validation?: Validation;
} & Base;
export type Length<T> = {
    max?: T;
    min?: T;
};